move method + error message
