+++ sdmetrics/single_table/data_augmentation/utils.py
@@ -2,7 +2,7 @@
 
 import pandas as pd
 
-from sdmetrics._utils_metadata import _validate_single_table_metadata
+from sdmetrics._utils_metadata import _process_data_with_metadata, _validate_single_table_metadata
 
 
 def _validate_tables(real_training_data, synthetic_data, real_validation_data):
@@ -126,3 +126,14 @@ def _validate_inputs(
         prediction_column_name,
         minority_class_label,
     )
+
+
+def _process_data_with_metadata_ml_efficacy_metrics(
+    real_training_data, synthetic_data, real_validation_data, metadata
+):
+    """Process the data for ML efficacy metrics according to the metadata."""
+    real_training_data = _process_data_with_metadata(real_training_data, metadata, True)
+    synthetic_data = _process_data_with_metadata(synthetic_data, metadata, True)
+    real_validation_data = _process_data_with_metadata(real_validation_data, metadata, True)
+
+    return real_training_data, synthetic_data, real_validation_data
